MQTT fix (resubscibe and republish after a reconnection) + lets devicePorts byte array instead of int to avoid frequent type casts

+++ cli/homeAssistant/messages.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 )
 
-func (ha *HomeAssistanceMqttClient) getDiscoveryMessage(devicePort int) (string, error) {
+func (ha *HomeAssistanceMqttClient) getDiscoveryMessage(devicePort byte) (string, error) {
 	name := ha.getUniqueObjectId(devicePort)
 	uniqueId := ha.getUniqueObjectId(devicePort)
 	device_class := `"device_class": "garage",`
